new lib for multiqc: MiltiqcSchema.groovy
